#365 chore : constant 폴더 domain 하위로 이동
