+++ src/main/java/com/floney/floney/analyze/service/AssetServiceImpl.java
@@ -19,9 +19,9 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-import static com.floney.floney.book.dto.constant.AssetType.BANK;
-import static com.floney.floney.book.dto.constant.AssetType.OUTCOME;
-import static com.floney.floney.book.dto.constant.CategoryEnum.FLOW;
+import static com.floney.floney.book.domain.constant.AssetType.BANK;
+import static com.floney.floney.book.domain.constant.AssetType.OUTCOME;
+import static com.floney.floney.book.domain.constant.CategoryEnum.FLOW;
 import static com.floney.floney.common.constant.Status.ACTIVE;
 
 @Service
