+++ src/main/java/com/floney/floney/book/domain/entity/CarryOver.java
@@ -1,6 +1,6 @@
 package com.floney.floney.book.domain.entity;
 
-import com.floney.floney.book.dto.constant.AssetType;
+import com.floney.floney.book.domain.constant.AssetType;
 import com.floney.floney.book.dto.request.BookLineRequest;
 import com.floney.floney.common.entity.BaseEntity;
 import com.querydsl.core.annotations.QueryProjection;
@@ -17,7 +17,7 @@ import javax.persistence.ManyToOne;
 import java.time.LocalDate;
 import java.util.Objects;
 
-import static com.floney.floney.book.dto.constant.AssetType.OUTCOME;
+import static com.floney.floney.book.domain.constant.AssetType.OUTCOME;
 
 @Entity
 @Getter
@@ -44,27 +44,27 @@ public class CarryOver extends BaseEntity {
     public static CarryOver of(BookLineRequest request, Book book, LocalDate date) {
         if (Objects.equals(request.getFlow(), OUTCOME.getKind())) {
             return CarryOver
-                    .builder()
-                    .money(-1 * request.getMoney())
-                    .book(book)
-                    .date(date)
-                    .build();
+                .builder()
+                .money(-1 * request.getMoney())
+                .book(book)
+                .date(date)
+                .build();
         } else {
             return CarryOver
-                    .builder()
-                    .money(request.getMoney())
-                    .book(book)
-                    .date(date)
-                    .build();
+                .builder()
+                .money(request.getMoney())
+                .book(book)
+                .date(date)
+                .build();
         }
 
     }
 
     public static CarryOver init() {
         return CarryOver
-                .builder()
-                .money(0L)
-                .build();
+            .builder()
+            .money(0L)
+            .build();
     }
 
     public void update(double updateMoney, String flow) {
@@ -76,11 +76,10 @@ public class CarryOver extends BaseEntity {
     }
 
     // 내역을 삭제하는 경우, 이월된 값을 되돌리기
-    public void delete(double updateMoney, BookLineCategory flow){
+    public void delete(double updateMoney, BookLineCategory flow) {
         if (Objects.equals(flow.getName(), AssetType.INCOME.getKind())) {
             money -= updateMoney;
-        }
-        else if (Objects.equals(flow.getName(), AssetType.OUTCOME.getKind())){
+        } else if (Objects.equals(flow.getName(), AssetType.OUTCOME.getKind())) {
             money += updateMoney;
         }
     }
