+++ src/main/java/com/floney/floney/book/dto/process/BookLineExpense.java
@@ -1,6 +1,6 @@
 package com.floney.floney.book.dto.process;
 
-import com.floney.floney.book.dto.constant.AssetType;
+import com.floney.floney.book.domain.constant.AssetType;
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
