+++ src/main/java/com/floney/floney/book/dto/request/ExcelDownloadRequest.java
@@ -1,6 +1,6 @@
 package com.floney.floney.book.dto.request;
 
-import com.floney.floney.book.dto.constant.ExcelDuration;
+import com.floney.floney.book.domain.constant.ExcelDuration;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
