+++ src/main/java/com/floney/floney/book/service/category/CategoryFactory.java
@@ -1,10 +1,10 @@
 package com.floney.floney.book.service.category;
 
-import com.floney.floney.book.dto.constant.CategoryEnum;
-import com.floney.floney.book.dto.request.BookLineRequest;
+import com.floney.floney.book.domain.constant.CategoryEnum;
 import com.floney.floney.book.domain.entity.BookLine;
 import com.floney.floney.book.domain.entity.BookLineCategory;
 import com.floney.floney.book.domain.entity.Category;
+import com.floney.floney.book.dto.request.BookLineRequest;
 import com.floney.floney.book.repository.category.BookLineCategoryRepository;
 import com.floney.floney.book.repository.category.CategoryRepository;
 import com.floney.floney.common.exception.book.NotFoundCategoryException;
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Component;
 import java.util.Map;
 import java.util.Objects;
 
-import static com.floney.floney.book.dto.constant.CategoryEnum.*;
+import static com.floney.floney.book.domain.constant.CategoryEnum.*;
 import static com.floney.floney.book.domain.entity.BookLineCategory.of;
 
 @Component
@@ -83,17 +83,17 @@ public class CategoryFactory {
 
     private Category findLineCategory(String line, String bookKey, String flow) {
         return categoryRepository.findLineCategory(line, bookKey, flow)
-                .orElseThrow(() -> new NotFoundCategoryException(line));
+            .orElseThrow(() -> new NotFoundCategoryException(line));
     }
 
     private Category findFlowCategory(String flow) {
         return categoryRepository.findFlowCategory(flow)
-                .orElseThrow(() -> new NotFoundCategoryException(flow));
+            .orElseThrow(() -> new NotFoundCategoryException(flow));
     }
 
     private Category findAssetCategory(String asset, String bookKey) {
         return categoryRepository.findAssetCategory(asset, bookKey)
-                .orElseThrow(() -> new NotFoundCategoryException(asset));
+            .orElseThrow(() -> new NotFoundCategoryException(asset));
     }
 
 }
